Consolidate parallelized autoresearch into tools.py and ui.py.
Merge agent tools, sandbox execution, memory helpers, and reporting into
two support modules so the tutorial directory stays focused on the entry
script, bundle, prepare, train, and shared types.
